Pause, Health, dying

+++ Assets/Code/Damage.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 public class Damage : MonoBehaviour
 {
-    [SerializeField] private float _damage = 1f;
+    [SerializeField] private int _damage = 1;
     void OnTriggerEnter2D(Collider2D other)
     {
         if (other.gameObject.GetComponent<Health>() != null)
